First version of initial and ending screen
Introduced methods for handling the creation of the initial and final scene to be displayed to the user;
Moved permission request at the top of onInit, so that it get asked before building the 3d scene;
Reintroduced on tap events, with few changes:
Removed looping the video, substituted with method for handling the playback end event.

For now the code shows two ways to display a message. Maybe we’ll find a better one.
The application is using a 360 photo as background. Maybe consider other options.