First attempt at updating the 3d scene according to the state of the application.
To be tested in depth. Ideally, a different scene can be created when there is no connection, or the permissions have not been granted, or the application has not received any intent, and the application should be able to distinguish every case with consistency.
The internet connection is not checked properly. The scenes are not created properly either.