Introduced internet checker, fixed handling scene creation, improved permission requests.
The application now checks the internet availability and the reachability of the mediaUri; also permissions for storage access are requested upfront if necessary; scene creation is handled properly for the main cases and when tapping the trackpad.
Removed multiple scene creation methods. Removed permission request and manager from Minimal360Video.