Enum to handle states and better thread synchronization
An enum is used to handle all the states of the application, states may be added in the future.
State changes after responses of asynchronous tasks are better handled.
Scene changes now happen correctly every time.
The app still crashes if no intent is provided and minor refactoring is needed