First skeleton of the replacement strategy:
The internal implementation (ExoPlayerImplInternal) hasn't been touched: in the method "maybeContinueLoading" the code queries the LoadControl (with the method shouldContinueLoading) for the buffer status: if the buffer has reached the maximum value (maxBufferMs) the method returns false, true otherwise.
In this first implementation, the method "shouldContinueLoading" return always true: the verification of the buffer is so delayed in the method "continueLoading" of the class DashSRDMediaPeriod.
In DashSRDMediaPeriod the following check is performed:
If the current buffered video duration minus the playback time (so the amount of video in the buffer) is more than the maxBufferMs and the video playback has started the chunk replacement method is called. The beginning of playback is checked because it does not make sense to replace chunks when the playback hasn't started because we have no information from the Picker on the user gaze. If the amount of video in the buffer is less then the maxBufferMs, the download continues to normally buffer chunks.

Add-ons
+ Picker information added
+ Buffer values passed
+ Safe Margin for the replacement added

ToDo:
Identify the chunk to be replaced using the temporal information (ms)
Identify a way to replace the chunk into the buffer
Handle the interruption of replacement function when we need to buffer more