Replacement in buffer attempt WIP
This is a first attempt at modifying the buffer underlying structure and allow replacements in it.
The core functionalities of the buffer seem to be implemented but some more classes from Exoplayer2 need to be copied and changed for it to work properly.
Thus the main buffer has not even been tested yet.