Introducing logging of freezing events
The application is now able to track when the playback freezes due to rebuffering. Each entry in the log stores a global time reference, the sequence number of the event, the duration of the freeze, and the playback time when the event occurred.
There might be some further testing to do. Also there seem to be some issue in the tiles picker that needs to be solved.