New buffer - one half decoded frame
Moving forward in the new buffer structure.
The buffer is managed using array lists instead of a fixed size byte array and fixed size info queue.
There are no exception raised in the current version and the first frame of the video is displayed (half decoded for some part of it).
Then the video stalls for some reason.
Needs to be fixed.