Added index counter to bandwidth tracker
From the bandwidth log file we can now know exactly how many files have been downloaded during the streaming process. Once the replacement strategy is introduced we can compute the percentage of chunks that have been downloaded twice.
Changed the starting index in the freeze log to reflect the fact that the first buffering event is not a rebuffering but it’s due to the startup delay.