fixed replacement condition and added opt out function
Replacement can now be deactivated by providing deactivateReplacement=true in the intent
Fixed the condition under which the replacement happens, the currently read chunck does not need to be in the buffer anymore.

Replacement cause weird chunk synchronization problems with the demo videos. The result is not good for viewing. Need investigations