Activated Pyramidal and search for the start and end time in the samples
Re-activated the Pyramidal strategy for the buffering.
Now search for the minimum and maximum times stored in the meta-data as the start and end time for the replacement.
The key frames the closest to those positions are identified in the buffer as the starting and ending position for the replacement.