More extensible XML parser
XML operations parser follows both a state pattern and a template method pattern.
States are used to determined what kind of object needs to be parsed for the incoming tags.
Template method determines one general process to follow to parse an XML element. Details of the parsing process then has to be specified for each new parsable operation element.